requirements.txt
Flask==1.1.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Werkzeug==0.15.5
numpy>=1.9.2
scipy>=0.15.1
scikit-learn>=0.18
matplotlib>=1.4.3
pandas>=0.19
Flask==1.1.2
imutils==0.5.3
numpy==1.18.1
opencv-python==4.2.0.32
pandas==1.0.1
Pillow==7.0.0
scikit-learn==0.22.1
scipy==1.3.2
tensorboard==2.0.0
tensorflow==2.0.0
tqdm==4.44.1
efficientnet==1.1.0
gevent==21.1.2
fpdf2==2.4.2